Refactor FileError type for FolderErrorsEventData.
Move into new package lib/model/types.
