+++ lib/model/folder.go
@@ -23,6 +23,7 @@ import (
 	"github.com/syncthing/syncthing/lib/fs"
 	"github.com/syncthing/syncthing/lib/ignore"
 	"github.com/syncthing/syncthing/lib/locations"
+	"github.com/syncthing/syncthing/lib/model/types"
 	"github.com/syncthing/syncthing/lib/osutil"
 	"github.com/syncthing/syncthing/lib/protocol"
 	"github.com/syncthing/syncthing/lib/scanner"
@@ -66,8 +67,8 @@ type folder struct {
 	pullPause     time.Duration
 	pullFailTimer *time.Timer
 
-	scanErrors []FileError
-	pullErrors []FileError
+	scanErrors []types.FileError
+	pullErrors []types.FileError
 	errorsMut  sync.Mutex
 
 	doInSyncChan chan syncRequest
@@ -1147,7 +1148,7 @@ func (f *folder) String() string {
 func (f *folder) newScanError(path string, err error) {
 	f.errorsMut.Lock()
 	l.Infof("Scanner (folder %s, item %q): %v", f.Description(), path, err)
-	f.scanErrors = append(f.scanErrors, FileError{
+	f.scanErrors = append(f.scanErrors, types.FileError{
 		Err:  err.Error(),
 		Path: path,
 	})
@@ -1174,11 +1175,11 @@ outer:
 	f.scanErrors = filtered
 }
 
-func (f *folder) Errors() []FileError {
+func (f *folder) Errors() []types.FileError {
 	f.errorsMut.Lock()
 	defer f.errorsMut.Unlock()
 	scanLen := len(f.scanErrors)
-	errors := make([]FileError, scanLen+len(f.pullErrors))
+	errors := make([]types.FileError, scanLen+len(f.pullErrors))
 	copy(errors[:scanLen], f.scanErrors)
 	copy(errors[scanLen:], f.pullErrors)
 	sort.Sort(fileErrorList(errors))
